Changed floats to doubles and improved data file loading

+++ src/main/scala/alankstewart/satin/Laser.scala
@@ -4,7 +4,7 @@ object CO2 extends Enumeration {
   val MD, PI = Value
 }
 
-class Laser(val outputFile: String, val smallSignalGain: Float, val dischargePressure: Int,
+class Laser(val outputFile: String, val smallSignalGain: Double, val dischargePressure: Int,
             val carbonDioxide: CO2.Value) {
 
   override def toString = outputFile + "  " + smallSignalGain + "  " + dischargePressure + "  " + carbonDioxide
